drop evil version-ing methods
