+++ lib/asciidoctor/rfc.rb
@@ -1,6 +1,6 @@
 require "asciidoctor"
 
 require "asciidoctor/rfc/version"
-require "asciidoctor/rfc/common"
+
 require "asciidoctor/rfc/v2/converter"
 require "asciidoctor/rfc/v3/converter"
