+++ lib/asciidoctor/rfc/v2/converter.rb
@@ -1,4 +1,4 @@
-require_relative "../common"
+require "asciidoctor/rfc/converter"
 
 module Asciidoctor
   module RFC::V2
@@ -11,6 +11,7 @@ module Asciidoctor
     class Converter
       include ::Asciidoctor::Converter
       include ::Asciidoctor::Writer
+      include ::Asciidoctor::RFC::Converter
 
       register_for "rfc2"
 
@@ -29,14 +30,6 @@ module Asciidoctor
         opts.empty? ? (send transform, node) : (send transform, node, opts)
       end
 
-      def content(node)
-        Asciidoctor::RFC::Common.content node
-      end
-
-      def skip(node, name = nil)
-        Asciidoctor::RFC::Common.skip node, name
-      end
-
       alias :pass :content
       alias :embedded :content
       alias :sidebar :content
@@ -51,14 +44,6 @@ module Asciidoctor
       alias :inline_menu :skip
       alias :verse :content
       alias :quote :content
-
-      def get_header_attribute(node, attr, default = nil)
-        Asciidoctor::RFC::Common.get_header_attribute node, attr, default
-      end
-
-      def set_header_attribute(attr, val)
-        Asciidoctor::RFC::Common.set_header_attribute attr, val
-      end
 
       def document_ns_attributes(_doc)
         # ' xmlns="http://projectmallard.org/1.0/" xmlns:its="http://www.w3.org/2005/11/its"'
@@ -145,6 +130,52 @@ module Asciidoctor
         result << (keyword node)
       end
 
+      def organization node, suffix
+        result = []
+        organization = node.attr("organization#{suffix}")
+        organization_abbrev = node.attr("organization_abbrev#{suffix}")
+        abbrev = set_header_attribute "abbrev", organization_abbrev
+        result << "<organization#{abbrev}>#{organization}</organization>" unless organization.nil?
+        result
+      end
+
+      def address node, suffix
+        result = []
+        postalline = nil
+        street = node.attr("street#{suffix}")
+        city = node.attr("city#{suffix}")
+        region = node.attr("region#{suffix}")
+        country = node.attr("country#{suffix}")
+        code = node.attr("code#{suffix}")
+        phone = node.attr("phone#{suffix}")
+        email = node.attr("email#{suffix}")
+        facsimile = node.attr("fax#{suffix}")
+        uri = node.attr("uri#{suffix}")
+        if (not email.nil?) || (not facsimile.nil?) || (not uri.nil?) || (not phone.nil?) ||
+          (not street.nil?)
+          result << "<address>"
+          if not street.nil?
+            result << "<postal>"
+            if postalline.nil?
+              street&.split("\\ ")&.each { |p| result << "<street>#{p}</street>" }
+              result << "<city>#{city}</city>" unless city.nil?
+              result << "<region>#{region}</region>" unless region.nil?
+              result << "<code>#{code}</code>" unless code.nil?
+              result << "<country>#{country}</country>" unless country.nil?
+            else
+              postalline&.split("\\ ")&.each { |p| result << "<postalLine>#{p}</postalLine>" }
+            end
+            result << "</postal>"
+          end
+          result << "<phone>#{phone}</phone>" unless phone.nil?
+          result << "<facsimile>#{facsimile}</facsimile>" unless facsimile.nil?
+          result << "<email>#{email}</email>" unless email.nil?
+          result << "<uri>#{uri}</uri>" unless uri.nil?
+          result << "</address>"
+        end
+        result
+      end
+
       def author(node)
         # = Title
         # Author;Author_2;Author_3
@@ -195,35 +226,56 @@ module Asciidoctor
         # :country
         # :code
         result = []
-        result << Asciidoctor::RFC::Common.authorname(node, suffix)
-        result << Asciidoctor::RFC::Common.organization(node, suffix, 2)
-        result << Asciidoctor::RFC::Common.address(node, suffix, 2)
+        result << authorname(node, suffix)
+        result << organization(node, suffix)
+        result << address(node, suffix)
         result << "</author>"
         result
       end
 
       def date node
-        Asciidoctor::RFC::Common.date(node, 2)
+        # = Title
+        # Author
+        # :revdate or :date
+        result = []
+        revdate = node.attr("revdate")
+        revdate = node.attr("date") if revdate.nil?
+        # date is mandatory in v2: use today
+        revdate = DateTime.now.iso8601 if revdate.nil?
+        warn %(asciidoctor: WARNING: revdate attribute missing from header, provided current date) 
+        unless revdate.nil?
+          begin
+            revdate.gsub!(/T.*$/, "")
+            d = Date.iso8601 revdate
+            day = set_header_attribute "day", d.day
+            month = set_header_attribute "month", Date::MONTHNAMES[d.month]
+            year = set_header_attribute "year", d.year
+            result << "<date#{day}#{month}#{year}/>"
+          rescue
+            # nop
+          end
+        end
+        result
       end
 
       def area node
-        Asciidoctor::RFC::Common.area(node)
+        area(node)
       end
 
       def workgroup node
-        Asciidoctor::RFC::Common.workgroup(node)
+        workgroup(node)
       end
 
       def keyword node
-        Asciidoctor::RFC::Common.keyword(node)
+        keyword(node)
       end
 
       def inline_anchor(node)
-        Asciidoctor::RFC::Common.inline_anchor(node)
+        inline_anchor(node)
       end
 
       def inline_indexterm(node)
-        Asciidoctor::RFC::Common.inline_indexterm(node)
+        inline_indexterm(node)
       end
 
       def inline_break(node)
@@ -328,7 +380,7 @@ module Asciidoctor
           end
           title = set_header_attribute "title", node.title
           result << "<note#{title}>"
-          result << (Asciidoctor::RFC::Common.paragraph1 node)
+          result << (paragraph1 node)
           result << "</note>"
         else
           id = set_header_attribute "anchor", node.id
@@ -487,7 +539,7 @@ module Asciidoctor
       end
 
       def listing(node)
-        Asciidoctor::RFC::Common.listing(node, 2)
+        listing(node, 2)
       end
 
       def ulist(node)
@@ -632,11 +684,65 @@ module Asciidoctor
       end
 
       def inline_image(node)
-        Asciidoctor::RFC::Common.inline_image(node, 2)
+        result = []
+        result << "<figure>" if node.parent.context != :example
+        align = get_header_attribute node, "align"
+        alt = get_header_attribute node, "alt"
+        link =  (node.image_uri node.target)
+        src = set_header_attribute node, "src", link
+        type = set_header_attribute node, "type", link =~ /\.svg$/ ? "svg" : "binary-art"
+        result << "<artwork#{align}#{alt}#{type}#{src}/>"
+        result << "</figure>" if node.parent.context != :example
+        result
       end
 
       def image(node)
-        Asciidoctor::RFC::Common.image(node, 2)
+        # [[id]]
+        # .Name
+        # [link=xxx,align=left|center|right,alt=alt_text,type]
+        # image::filename[]
+        # ignoring width, height attributes
+        result = []
+        result << "<figure>" if node.parent.context != :example
+        id = set_header_attribute "anchor", node.id
+        align = get_header_attribute node, "align"
+        alt = set_header_attribute "alt", node.alt
+        link = (node.image_uri node.target)
+        src = set_header_attribute node, "src", link
+        type = get_header_attribute node, "type"
+        name = get_header_attribute node, "name"
+        result << "<artwork#{id}#{name}#{align}#{alt}#{type}#{src}/>"
+        result << "</figure>" if node.parent.context != :example
+        result
+      end
+
+      def listing node
+=begin
+.name
+[source,type,src=uri] (src is mutually exclusive with listing content) (v3)
+[source,type,src=uri,align,alt] (src is mutually exclusive with listing content) (v2)
+----
+code
+----
+=end
+        result = []
+        result << "<figure>" if node.parent.context != :example
+        align = set_header_attribute "align", node.title
+        alt = set_header_attribute "alt", node.alt
+        tag = "artwork"
+        id = set_header_attribute "anchor", node.id
+        name = set_header_attribute "name", node.title
+        type = set_header_attribute "type", node.attr("language")
+        src = set_header_attribute "src", node.attr("src")
+        result << "<#{tag}#{id}#{align}#{name}#{type}#{src}#{alt}>"
+        if src.nil?
+          node.lines.each do |line|
+            result << line.gsub(/\&/, "&amp;").gsub(/</, "&lt;").gsub(/>/, "&gt;")
+          end
+        end
+        result << "</#{tag}>"
+        result << "</figure>" if node.parent.context != :example
+        result
       end
     end
   end
